refactor : 기존의 MVC패턴으로 수정하고, Game은 controllers, Referee는 models로 옮김
- data 폴더명 -> models로 수정
- domain  폴더 삭제
