+++ src/controllers/index.js
@@ -1,3 +1,4 @@
+import Game from './Game.js';
 import InputController from './InputController.js';
 
-export { InputController };
+export { Game,InputController };
