--- src/domains/index.js
@@ -1,4 +1,0 @@
-import Game from './Game.js';
-import Referee from './Referee.js';
-
-export { Game, Referee };
