+++ src/models/index.js
@@ -1,5 +1,6 @@
 import Car from './Car.js';
 import CarList from './CarList.js';
+import Referee from './Referee.js';
 import Round from './Round.js';
 
-export { Car, CarList, Round };
+export { Car, CarList, Referee, Round };
