+++ src/view/index.js
@@ -1,3 +1,4 @@
 import InputView from './InputView.js';
 import OutputView from './OutputView.js';
+
 export { InputView, OutputView };
